refactor: 重构页面骨架第一弹
