+++ web/src/lib/components/app-sidebar.svelte
@@ -1,226 +1,150 @@
-<script lang="ts">
-	import ChevronRightIcon from '@lucide/svelte/icons/chevron-right';
-	import SettingsIcon from '@lucide/svelte/icons/settings';
-	import UserIcon from '@lucide/svelte/icons/user';
-	import HeartIcon from '@lucide/svelte/icons/heart';
-	import FolderIcon from '@lucide/svelte/icons/folder';
+<script lang="ts" module>
+	import BookOpenIcon from '@lucide/svelte/icons/book-open';
 	import DatabaseIcon from '@lucide/svelte/icons/database';
-	import * as Sidebar from '$lib/components/ui/sidebar/index.js';
-	import { useSidebar } from '$lib/components/ui/sidebar/context.svelte.js';
-	import {
-		appStateStore,
-		setVideoSourceFilter,
-		clearAll,
-		ToQuery,
-		resetCurrentPage
-	} from '$lib/stores/filter';
+	import FileVideoIcon from '@lucide/svelte/icons/file-video';
+	import BotIcon from '@lucide/svelte/icons/bot';
+	import ChartPieIcon from '@lucide/svelte/icons/chart-pie';
+	import LifeBuoyIcon from '@lucide/svelte/icons/life-buoy';
+	import MapIcon from '@lucide/svelte/icons/map';
+	import SendIcon from '@lucide/svelte/icons/send';
+	import Settings2Icon from '@lucide/svelte/icons/settings-2';
+	import SquareTerminalIcon from '@lucide/svelte/icons/square-terminal';
 
-	import { type VideoSourcesResponse } from '$lib/types';
-	import { VIDEO_SOURCES } from '$lib/consts';
-	import * as Collapsible from '$lib/components/ui/collapsible/index.js';
-	import { goto } from '$app/navigation';
-	import { videoSourceStore } from '$lib/stores/video-source';
-	const sidebar = useSidebar();
-
-	const items = Object.values(VIDEO_SOURCES);
-
-	function handleSourceClick(sourceType: string, sourceId: number) {
-		setVideoSourceFilter({
-			type: sourceType,
-			id: sourceId.toString()
-		});
-		resetCurrentPage();
-		goto(`/${ToQuery($appStateStore)}`);
-		if (sidebar.isMobile) {
-			sidebar.setOpenMobile(false);
-		}
-	}
-
-	function handleLogoClick() {
-		clearAll();
-		goto('/');
-
-		if (sidebar.isMobile) {
-			sidebar.setOpenMobile(false);
-		}
-	}
+	const data = {
+		header: {
+			title: 'Bili Sync',
+			subtitle: '后台管理系统',
+			icon: BotIcon,
+			href: '/'
+		},
+		navMain: [
+			{
+				category: '总览',
+				items: [
+					{
+						title: '仪表盘',
+						icon: ChartPieIcon,
+						href: '/'
+					},
+					{
+						title: '日志',
+						icon: SquareTerminalIcon,
+						href: '/logs'
+					}
+				]
+			},
+			{
+				category: '内容管理',
+				items: [
+					{
+						title: '视频',
+						icon: FileVideoIcon,
+						href: '/videos'
+					},
+					{
+						title: '视频源',
+						icon: DatabaseIcon,
+						href: '/video-sources'
+					}
+				]
+			},
+			{
+				category: '快捷订阅',
+				items: [
+					{
+						title: '我创建的收藏夹',
+						icon: Settings2Icon,
+						href: '/me/favorites'
+					},
+					{
+						title: '我关注的合集',
+						icon: MapIcon,
+						href: '/me/collections'
+					},
+					{
+						title: '我关注的 up 主',
+						icon: LifeBuoyIcon,
+						href: '/me/uppers'
+					}
+				]
+			}
+		],
+		footer: [
+			{
+				title: '设置',
+				icon: Settings2Icon,
+				href: '/settings'
+			}
+		]
+	};
 </script>
 
-<Sidebar.Root class="border-border bg-background border-r">
-	<Sidebar.Header class="border-border flex h-[73px] items-center border-b">
-		<a
-			href="/"
-			class="flex w-full items-center gap-3 px-4 py-3 hover:cursor-pointer"
-			onclick={handleLogoClick}
-		>
-			<div class="flex h-8 w-8 items-center justify-center overflow-hidden rounded-lg">
-				<img src="/favicon.png" alt="Bili Sync" class="h-6 w-6" />
-			</div>
-			<div class="grid flex-1 text-left text-sm leading-tight">
-				<span class="truncate font-semibold">Bili Sync</span>
-				<span class="text-muted-foreground truncate text-xs">视频管理系统</span>
-			</div>
-		</a>
+<script lang="ts">
+	import * as Sidebar from '$lib/components/ui/sidebar/index.js';
+	import CommandIcon from '@lucide/svelte/icons/command';
+	import type { ComponentProps } from 'svelte';
+
+	let { ref = $bindable(null), ...restProps }: ComponentProps<typeof Sidebar.Root> = $props();
+</script>
+
+<Sidebar.Root bind:ref variant="inset" {...restProps}>
+	<Sidebar.Header>
+		<Sidebar.Menu>
+			<Sidebar.MenuItem>
+				<Sidebar.MenuButton size="lg">
+					{#snippet child({ props })}
+						<a href={data.header.href} {...props}>
+							<div
+								class="bg-sidebar-primary text-sidebar-primary-foreground flex aspect-square size-8 items-center justify-center rounded-lg"
+							>
+								<data.header.icon class="size-4" />
+							</div>
+							<div class="grid flex-1 text-left text-sm leading-tight">
+								<span class="truncate font-medium">{data.header.title}</span>
+								<span class="truncate text-xs">{data.header.subtitle}</span>
+							</div>
+						</a>
+					{/snippet}
+				</Sidebar.MenuButton>
+			</Sidebar.MenuItem>
+		</Sidebar.Menu>
 	</Sidebar.Header>
-	<Sidebar.Content class="flex flex-col px-2 py-3">
-		<div class="flex-1">
-			<Sidebar.Group>
-				<Sidebar.GroupLabel
-					class="text-muted-foreground mb-2 px-2 text-xs font-medium tracking-wider uppercase"
-				>
-					视频筛选
-				</Sidebar.GroupLabel>
-				<Sidebar.GroupContent>
-					<Sidebar.Menu class="space-y-1">
-						{#each items as item (item.type)}
-							<Collapsible.Root class="group/collapsible">
-								<Sidebar.MenuItem>
-									<Collapsible.Trigger class="w-full">
-										{#snippet child({ props })}
-											<Sidebar.MenuButton
-												{...props}
-												class="hover:bg-accent/50 text-foreground flex w-full cursor-pointer items-center justify-between rounded-lg px-3 py-2.5 font-medium transition-all duration-200"
-											>
-												<div class="flex flex-1 items-center gap-3">
-													<item.icon class="text-muted-foreground h-4 w-4" />
-													<span class="text-sm">{item.title}</span>
-												</div>
-												<ChevronRightIcon
-													class="text-muted-foreground h-3 w-3 transition-transform duration-200 group-data-[state=open]/collapsible:rotate-90"
-												/>
-											</Sidebar.MenuButton>
-										{/snippet}
-									</Collapsible.Trigger>
-									<Collapsible.Content class="mt-1">
-										<div class="border-border ml-5 space-y-0.5 border-l pl-2">
-											{#if $videoSourceStore}
-												{#if $videoSourceStore[item.type as keyof VideoSourcesResponse]?.length > 0}
-													{#each $videoSourceStore[item.type as keyof VideoSourcesResponse] as source (source.id)}
-														<Sidebar.MenuItem>
-															<button
-																class="text-foreground hover:bg-accent/50 w-full cursor-pointer rounded-md px-3 py-2 text-left text-sm transition-all duration-200"
-																onclick={() => handleSourceClick(item.type, source.id)}
-															>
-																<span class="block truncate">{source.name}</span>
-															</button>
-														</Sidebar.MenuItem>
-													{/each}
-												{:else}
-													<div class="text-muted-foreground px-3 py-2 text-sm">无数据</div>
-												{/if}
-											{:else}
-												<div class="text-muted-foreground px-3 py-2 text-sm">加载中...</div>
-											{/if}
-										</div>
-									</Collapsible.Content>
-								</Sidebar.MenuItem>
-							</Collapsible.Root>
-						{/each}
-					</Sidebar.Menu>
-				</Sidebar.GroupContent>
-			</Sidebar.Group>
-
-			<Sidebar.Group>
-				<Sidebar.GroupLabel
-					class="text-muted-foreground mb-2 px-2 text-xs font-medium tracking-wider uppercase"
-				>
-					快捷订阅
-				</Sidebar.GroupLabel>
-				<Sidebar.GroupContent>
-					<Sidebar.Menu class="space-y-1">
+	<Sidebar.Content>
+		<Sidebar.Group>
+			{#each data.navMain as group (group.category)}
+				<Sidebar.GroupLabel class="h-10">{group.category}</Sidebar.GroupLabel>
+				<Sidebar.Menu>
+					{#each group.items as item (item.title)}
 						<Sidebar.MenuItem>
-							<Sidebar.MenuButton>
-								<a
-									href="/me/favorites"
-									class="hover:bg-accent/50 text-foreground flex w-full cursor-pointer items-center gap-3 rounded-lg px-3 py-2.5 text-sm font-medium transition-all duration-200"
-									onclick={() => {
-										if (sidebar.isMobile) {
-											sidebar.setOpenMobile(false);
-										}
-									}}
-								>
-									<HeartIcon class="text-muted-foreground h-4 w-4" />
-									<span>创建的收藏夹</span>
-								</a>
+							<Sidebar.MenuButton class="h-8">
+								{#snippet child({ props })}
+									<a href={item.href} {...props}>
+										<item.icon class="size-4" />
+										<span class="text-sm">{item.title}</span>
+									</a>
+								{/snippet}
 							</Sidebar.MenuButton>
 						</Sidebar.MenuItem>
-						<Sidebar.MenuItem>
-							<Sidebar.MenuButton>
-								<a
-									href="/me/collections"
-									class="hover:bg-accent/50 text-foreground flex w-full cursor-pointer items-center gap-3 rounded-lg px-3 py-2.5 text-sm font-medium transition-all duration-200"
-									onclick={() => {
-										if (sidebar.isMobile) {
-											sidebar.setOpenMobile(false);
-										}
-									}}
-								>
-									<FolderIcon class="text-muted-foreground h-4 w-4" />
-									<span>关注的合集</span>
-								</a>
-							</Sidebar.MenuButton>
-						</Sidebar.MenuItem>
-						<Sidebar.MenuItem>
-							<Sidebar.MenuButton>
-								<a
-									href="/me/uppers"
-									class="hover:bg-accent/50 text-foreground flex w-full cursor-pointer items-center gap-3 rounded-lg px-3 py-2.5 text-sm font-medium transition-all duration-200"
-									onclick={() => {
-										if (sidebar.isMobile) {
-											sidebar.setOpenMobile(false);
-										}
-									}}
-								>
-									<UserIcon class="text-muted-foreground h-4 w-4" />
-									<span>关注的 UP 主</span>
-								</a>
-							</Sidebar.MenuButton>
-						</Sidebar.MenuItem>
-					</Sidebar.Menu>
-				</Sidebar.GroupContent>
-			</Sidebar.Group>
-		</div>
-
-		<!-- 固定在底部的菜单选项 -->
-		<div class="border-border mt-auto border-t pt-4">
-			<Sidebar.Menu class="space-y-1">
-				<Sidebar.MenuItem>
-					<Sidebar.MenuButton>
-						<a
-							href="/video-sources"
-							class="hover:bg-accent/50 text-foreground flex w-full cursor-pointer items-center justify-between rounded-lg px-3 py-2.5 font-medium transition-all duration-200"
-							onclick={() => {
-								if (sidebar.isMobile) {
-									sidebar.setOpenMobile(false);
-								}
-							}}
-						>
-							<div class="flex flex-1 items-center gap-3">
-								<DatabaseIcon class="text-muted-foreground h-4 w-4" />
-								<span class="text-sm">视频源管理</span>
-							</div>
-						</a>
-					</Sidebar.MenuButton>
-				</Sidebar.MenuItem>
-				<Sidebar.MenuItem>
-					<Sidebar.MenuButton>
-						<a
-							href="/settings"
-							class="hover:bg-accent/50 text-foreground flex w-full cursor-pointer items-center justify-between rounded-lg px-3 py-2.5 font-medium transition-all duration-200"
-							onclick={() => {
-								if (sidebar.isMobile) {
-									sidebar.setOpenMobile(false);
-								}
-							}}
-						>
-							<div class="flex flex-1 items-center gap-3">
-								<SettingsIcon class="text-muted-foreground h-4 w-4" />
-								<span class="text-sm">设置</span>
-							</div>
-						</a>
-					</Sidebar.MenuButton>
-				</Sidebar.MenuItem>
-			</Sidebar.Menu>
-		</div>
+					{/each}
+				</Sidebar.Menu>
+			{/each}
+		</Sidebar.Group>
 	</Sidebar.Content>
+	<Sidebar.Footer>
+		<Sidebar.Separator />
+		<Sidebar.Menu>
+			{#each data.footer as item (item.title)}
+				<Sidebar.MenuItem>
+					<Sidebar.MenuButton class="h-8">
+						{#snippet child({ props })}
+							<a href={item.href} {...props}>
+								<item.icon class="size-4" />
+								<span class="text-sm">{item.title}</span>
+							</a>
+						{/snippet}
+					</Sidebar.MenuButton>
+				</Sidebar.MenuItem>
+			{/each}
+		</Sidebar.Menu>
+	</Sidebar.Footer>
 </Sidebar.Root>
