+++ web/src/lib/components/ui/collapsible/collapsible.svelte
@@ -8,4 +8,4 @@
 	}: CollapsiblePrimitive.RootProps = $props();
 </script>
 
-<CollapsiblePrimitive.Root bind:ref data-slot="collapsible" {...restProps} />
+<CollapsiblePrimitive.Root bind:ref bind:open data-slot="collapsible" {...restProps} />
