+++ web/src/lib/components/ui/sheet/sheet-overlay.svelte
@@ -7,8 +7,6 @@
 		class: className,
 		...restProps
 	}: SheetPrimitive.OverlayProps = $props();
-
-	export { className as class };
 </script>
 
 <SheetPrimitive.Overlay
