+++ web/src/lib/hooks/is-mobile.svelte.ts
@@ -1,9 +1,9 @@
 import { MediaQuery } from 'svelte/reactivity';
 
-const MOBILE_BREAKPOINT = 768;
+const DEFAULT_MOBILE_BREAKPOINT = 768;
 
 export class IsMobile extends MediaQuery {
-	constructor() {
-		super(`max-width: ${MOBILE_BREAKPOINT - 1}px`);
+	constructor(breakpoint: number = DEFAULT_MOBILE_BREAKPOINT) {
+		super(`max-width: ${breakpoint - 1}px`);
 	}
 }
