+++ web/src/routes/+page.svelte
@@ -1,90 +1,51 @@
 <script lang="ts">
-	import VideoCard from '$lib/components/video-card.svelte';
-	import Pagination from '$lib/components/pagination.svelte';
-	import { Button } from '$lib/components/ui/button/index.js';
-	import * as AlertDialog from '$lib/components/ui/alert-dialog/index.js';
-	import RotateCcwIcon from '@lucide/svelte/icons/rotate-ccw';
-	import api from '$lib/api';
-	import type { VideosResponse, VideoSourcesResponse, ApiError } from '$lib/types';
-	import { onMount } from 'svelte';
-	import { page } from '$app/stores';
-	import HeartIcon from '@lucide/svelte/icons/heart';
-	import { goto } from '$app/navigation';
-	import { videoSourceStore } from '$lib/stores/video-source';
-	import { VIDEO_SOURCES } from '$lib/consts';
+	import { onMount, onDestroy } from 'svelte';
+	import { Card, CardContent, CardHeader, CardTitle } from '$lib/components/ui/card/index.js';
+	import { Progress } from '$lib/components/ui/progress/index.js';
+	import { Badge } from '$lib/components/ui/badge/index.js';
+	import * as Chart from '$lib/components/ui/chart/index.js';
+	import MyChartTooltip from '$lib/components/custom/my-chart-tooltip.svelte';
+	import { curveNatural } from 'd3-shape';
+	import { BarChart, AreaChart } from 'layerchart';
 	import { setBreadcrumb } from '$lib/stores/breadcrumb';
-	import {
-		appStateStore,
-		clearVideoSourceFilter,
-		resetCurrentPage,
-		setAll,
-		setCurrentPage,
-		ToQuery
-	} from '$lib/stores/filter';
 	import { toast } from 'svelte-sonner';
-	import DropdownFilter from '$lib/components/dropdown-filter.svelte';
+	import api from '$lib/api';
+	import type { DashBoardResponse, SysInfoResponse, ApiError } from '$lib/types';
+	import DatabaseIcon from '@lucide/svelte/icons/database';
+	import HeartIcon from '@lucide/svelte/icons/heart';
+	import FolderIcon from '@lucide/svelte/icons/folder';
+	import UserIcon from '@lucide/svelte/icons/user';
+	import ClockIcon from '@lucide/svelte/icons/clock';
+	import VideoIcon from '@lucide/svelte/icons/video';
+	import HardDriveIcon from '@lucide/svelte/icons/hard-drive';
+	import CpuIcon from '@lucide/svelte/icons/cpu';
+	import MemoryStickIcon from '@lucide/svelte/icons/memory-stick';
 
-	const pageSize = 20;
-
-	let videosData: VideosResponse | null = null;
+	let dashboardData: DashBoardResponse | null = null;
+	let sysInfo: SysInfoResponse | null = null;
 	let loading = false;
+	let sysInfoEventSource: EventSource | null = null;
 
-	let lastSearch: string | null = null;
-
-	let resetAllDialogOpen = false;
-	let resettingAll = false;
-
-	function getApiParams(searchParams: URLSearchParams) {
-		let videoSource = null;
-		for (const source of Object.values(VIDEO_SOURCES)) {
-			const value = searchParams.get(source.type);
-			if (value) {
-				videoSource = { type: source.type, id: value };
-			}
-		}
-		return {
-			query: searchParams.get('query') || '',
-			videoSource,
-			pageNum: parseInt(searchParams.get('page') || '0')
-		};
+	function formatBytes(bytes: number): string {
+		if (bytes === 0) return '0 B';
+		const k = 1024;
+		const sizes = ['B', 'KB', 'MB', 'GB', 'TB'];
+		const i = Math.floor(Math.log(bytes) / Math.log(k));
+		return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i];
 	}
 
-	function getFilterContent(type: string, id: string) {
-		const filterTitle = Object.values(VIDEO_SOURCES).find((s) => s.type === type)?.title || '';
-		let filterName = '';
-		const videoSources = $videoSourceStore;
-		if (videoSources && type && id) {
-			const sources = videoSources[type as keyof VideoSourcesResponse];
-			filterName = sources?.find((s) => s.id.toString() === id)?.name || '';
-		}
-		return {
-			title: filterTitle,
-			name: filterName
-		};
+	function formatCpu(cpu: number): string {
+		return `${cpu.toFixed(1)}%`;
 	}
 
-	async function loadVideos(
-		query: string,
-		pageNum: number = 0,
-		filter?: { type: string; id: string } | null
-	) {
+	async function loadDashboard() {
 		loading = true;
 		try {
-			const params: Record<string, string | number> = {
-				page: pageNum,
-				page_size: pageSize
-			};
-			if (query) {
-				params.query = query;
-			}
-			if (filter) {
-				params[filter.type] = parseInt(filter.id);
-			}
-			const result = await api.getVideos(params);
-			videosData = result.data;
+			const response = await api.getDashboard();
+			dashboardData = response.data;
 		} catch (error) {
-			console.error('加载视频失败:', error);
-			toast.error('加载视频失败', {
+			console.error('加载仪表盘数据失败:', error);
+			toast.error('加载仪表盘数据失败', {
 				description: (error as ApiError).message
 			});
 		} finally {
@@ -92,208 +53,373 @@
 		}
 	}
 
-	async function handlePageChange(pageNum: number) {
-		setCurrentPage(pageNum);
-		goto(`/${ToQuery($appStateStore)}`);
-	}
-
-	async function handleSearchParamsChange(searchParams: URLSearchParams) {
-		const { query, videoSource, pageNum } = getApiParams(searchParams);
-		setAll(query, pageNum, videoSource);
-		loadVideos(query, pageNum, videoSource);
-	}
-
-	function handleFilterRemove() {
-		clearVideoSourceFilter();
-		resetCurrentPage();
-		goto(`/${ToQuery($appStateStore)}`);
-	}
-
-	async function handleResetVideo(id: number) {
+	// 启动系统信息流
+	function startSysInfoStream() {
 		try {
-			const result = await api.resetVideo(id);
-			const data = result.data;
-			if (data.resetted) {
-				toast.success('重置成功', {
-					description: `视频「${data.video.name}」已重置`
-				});
-				const { query, currentPage, videoSource } = $appStateStore;
-				await loadVideos(query, currentPage, videoSource);
-			} else {
-				toast.info('重置无效', {
-					description: `视频「${data.video.name}」没有失败的状态，无需重置`
-				});
-			}
+			sysInfoEventSource = api.createSysInfoStream(
+				(data) => {
+					sysInfo = data;
+				},
+				(_error) => {
+					toast.error('系统信息流异常中断');
+				}
+			);
 		} catch (error) {
-			console.error('重置失败:', error);
-			toast.error('重置失败', {
-				description: (error as ApiError).message
-			});
+			console.error('启动系统信息流失败:', error);
 		}
 	}
 
-	async function handleResetAllVideos() {
-		resettingAll = true;
-		try {
-			const result = await api.resetAllVideos();
-			const data = result.data;
-			if (data.resetted) {
-				toast.success('重置成功', {
-					description: `已重置 ${data.resetted_videos_count} 个视频和 ${data.resetted_pages_count} 个分页`
-				});
-				const { query, currentPage, videoSource } = $appStateStore;
-				await loadVideos(query, currentPage, videoSource);
-			} else {
-				toast.info('没有需要重置的视频');
-			}
-		} catch (error) {
-			console.error('重置失败:', error);
-			toast.error('重置失败', {
-				description: (error as ApiError).message
-			});
-		} finally {
-			resettingAll = false;
-			resetAllDialogOpen = false;
+	// 停止系统信息流
+	function stopSysInfoStream() {
+		if (sysInfoEventSource) {
+			sysInfoEventSource.close();
+			sysInfoEventSource = null;
 		}
 	}
 
-	$: if ($page.url.search !== lastSearch) {
-		lastSearch = $page.url.search;
-		handleSearchParamsChange($page.url.searchParams);
-	}
-
-	onMount(async () => {
-		setBreadcrumb([
-			{
-				label: '主页',
-				isActive: true
-			}
-		]);
+	onMount(() => {
+		setBreadcrumb([{ label: '仪表盘', isActive: true }]);
+		loadDashboard();
+		startSysInfoStream();
 	});
 
-	$: totalPages = videosData ? Math.ceil(videosData.total_count / pageSize) : 0;
-	$: filterContent = $appStateStore.videoSource
-		? getFilterContent($appStateStore.videoSource.type, $appStateStore.videoSource.id)
-		: { title: '', name: '' };
+	onDestroy(() => {
+		stopSysInfoStream();
+	});
+
+	// 图表配置
+	const videoChartConfig = {
+		videos: {
+			label: '视频数量',
+			color: 'var(--chart-1)'
+		}
+	} satisfies Chart.ChartConfig;
+
+	const memoryChartConfig = {
+		used: {
+			label: '整体占用',
+			color: 'var(--chart-1)'
+		},
+		process: {
+			label: '程序占用',
+			color: 'var(--chart-2)'
+		}
+	} satisfies Chart.ChartConfig;
+
+	const cpuChartConfig = {
+		used: {
+			label: '整体占用',
+			color: 'var(--chart-1)'
+		},
+		process: {
+			label: '程序占用',
+			color: 'var(--chart-2)'
+		}
+	} satisfies Chart.ChartConfig;
+
+	// 内存和 CPU 数据历史记录
+	let memoryHistory: Array<{ time: Date; used: number; process: number }> = [];
+	let cpuHistory: Array<{ time: Date; used: number; process: number }> = [];
+
+	// 更新历史数据
+	$: if (sysInfo) {
+		memoryHistory = [
+			...memoryHistory.slice(-19),
+			{
+				time: new Date(),
+				used: sysInfo.used_memory,
+				process: sysInfo.process_memory
+			}
+		];
+		cpuHistory = [
+			...cpuHistory.slice(-19),
+			{
+				time: new Date(),
+				used: sysInfo.used_cpu,
+				process: sysInfo.process_cpu
+			}
+		];
+	}
+	// 计算磁盘使用率
+	$: diskUsagePercent = sysInfo
+		? ((sysInfo.total_disk - sysInfo.available_disk) / sysInfo.total_disk) * 100
+		: 0;
 </script>
 
 <svelte:head>
-	<title>主页 - Bili Sync</title>
+	<title>仪表盘 - Bili Sync</title>
 </svelte:head>
 
-<div class="mb-4 flex items-center gap-2">
-	<span class="text-muted-foreground text-sm">当前筛选:</span>
-	<DropdownFilter
-		title="筛选"
-		filters={[
-			{
-				key: 'videoSource',
-				name: '视频来源',
-				icon: HeartIcon,
-				values: [
-					{
-						name: '全部',
-						id: ''
-					}
-				]
-			}
-		]}
-		selectedLabel={{
-			key: 'videoSource',
-			name: '视频来源',
-			valueName: filterContent.name || '全部',
-			valueId: $appStateStore.videoSource?.id || ''
-		}}
-	/>
+<div class="space-y-6">
+	{#if loading}
+		<div class="flex items-center justify-center py-12">
+			<div class="text-muted-foreground">加载中...</div>
+		</div>
+	{:else}
+		<div class="grid gap-4 md:grid-cols-3">
+			<!-- 存储空间卡片 -->
+			<Card class="md:col-span-1">
+				<CardHeader class="flex flex-row items-center justify-between space-y-0 pb-2">
+					<CardTitle class="text-sm font-medium">存储空间</CardTitle>
+					<HardDriveIcon class="text-muted-foreground h-4 w-4" />
+				</CardHeader>
+				<CardContent>
+					{#if sysInfo}
+						<div class="space-y-2">
+							<div class="flex items-center justify-between">
+								<div class="text-2xl font-bold">{formatBytes(sysInfo.available_disk)} 可用</div>
+								<div class="text-muted-foreground text-sm">
+									共 {formatBytes(sysInfo.total_disk)}
+								</div>
+							</div>
+							<Progress value={diskUsagePercent} class="h-2" />
+							<div class="text-muted-foreground text-xs">
+								已使用 {diskUsagePercent.toFixed(1)}% 的存储空间
+							</div>
+						</div>
+					{:else}
+						<div class="text-muted-foreground text-sm">加载中...</div>
+					{/if}
+				</CardContent>
+			</Card>
+			<Card class="md:col-span-2">
+				<CardHeader class="flex flex-row items-center justify-between space-y-0 pb-2">
+					<CardTitle class="text-sm font-medium">当前监听</CardTitle>
+					<DatabaseIcon class="text-muted-foreground h-4 w-4" />
+				</CardHeader>
+				<CardContent>
+					{#if dashboardData}
+						<div class="grid grid-cols-2 gap-4">
+							<div class="flex items-center justify-between">
+								<div class="flex items-center gap-2">
+									<HeartIcon class="text-muted-foreground h-4 w-4" />
+									<span class="text-sm">收藏夹</span>
+								</div>
+								<Badge variant="outline">{dashboardData.enabled_favorites}</Badge>
+							</div>
+							<div class="flex items-center justify-between">
+								<div class="flex items-center gap-2">
+									<FolderIcon class="text-muted-foreground h-4 w-4" />
+									<span class="text-sm">合集</span>
+								</div>
+								<Badge variant="outline">{dashboardData.enabled_collections}</Badge>
+							</div>
+							<div class="flex items-center justify-between">
+								<div class="flex items-center gap-2">
+									<UserIcon class="text-muted-foreground h-4 w-4" />
+									<span class="text-sm">投稿</span>
+								</div>
+								<Badge variant="outline">{dashboardData.enabled_submissions}</Badge>
+							</div>
+							<div class="flex items-center justify-between">
+								<div class="flex items-center gap-2">
+									<ClockIcon class="text-muted-foreground h-4 w-4" />
+									<span class="text-sm">稍后再看</span>
+								</div>
+								<Badge variant="outline">
+									{dashboardData.enable_watch_later ? '启用' : '禁用'}
+								</Badge>
+							</div>
+						</div>
+					{:else}
+						<div class="text-muted-foreground text-sm">加载中...</div>
+					{/if}
+				</CardContent>
+			</Card>
+		</div>
+
+		<div class="grid grid-cols-1 gap-4">
+			<Card>
+				<CardHeader class="flex flex-row items-center justify-between space-y-0 pb-2">
+					<CardTitle class="text-sm font-medium">最近入库</CardTitle>
+					<VideoIcon class="text-muted-foreground h-4 w-4" />
+				</CardHeader>
+				<CardContent>
+					{#if dashboardData && dashboardData.videos_by_day.length > 0}
+						<div class="mb-4 space-y-2">
+							<div class="flex items-center justify-between text-sm">
+								<span>近七日共新增视频</span>
+								<span class="font-medium"
+									>{dashboardData.videos_by_day.reduce((sum, v) => sum + v.cnt, 0)} 个</span
+								>
+							</div>
+						</div>
+						<Chart.Container config={videoChartConfig} class="h-[200px] w-full">
+							<BarChart
+								data={dashboardData.videos_by_day}
+								x="day"
+								axis="x"
+								series={[
+									{
+										key: 'cnt',
+										label: '新增视频',
+										color: videoChartConfig.videos.color
+									}
+								]}
+								props={{
+									bars: {
+										stroke: 'none',
+										rounded: 'all',
+										radius: 8,
+										initialHeight: 0
+									},
+									highlight: { area: { fill: 'none' } },
+									xAxis: { format: () => '' }
+								}}
+							>
+								{#snippet tooltip()}
+									<MyChartTooltip indicator="line" />
+								{/snippet}
+							</BarChart>
+						</Chart.Container>
+					{:else}
+						<div class="text-muted-foreground flex h-[300px] items-center justify-center text-sm">
+							暂无视频统计数据
+						</div>
+					{/if}</CardContent
+				>
+			</Card>
+		</div>
+
+		<!-- 第三行：系统监控 -->
+		<div class="grid gap-4 md:grid-cols-2">
+			<!-- 内存使用情况 -->
+			<Card>
+				<CardHeader class="flex flex-row items-center justify-between space-y-0 pb-2">
+					<CardTitle class="text-sm font-medium">内存使用情况</CardTitle>
+					<MemoryStickIcon class="text-muted-foreground h-4 w-4" />
+				</CardHeader>
+				<CardContent>
+					{#if sysInfo}
+						<div class="mb-4 space-y-2">
+							<div class="flex items-center justify-between text-sm">
+								<span>当前内存使用</span>
+								<span class="font-medium"
+									>{formatBytes(sysInfo.used_memory)} / {formatBytes(sysInfo.total_memory)}</span
+								>
+							</div>
+						</div>
+					{/if}
+					{#if memoryHistory.length > 0}
+						<Chart.Container config={memoryChartConfig} class="h-[150px] w-full">
+							<AreaChart
+								data={memoryHistory}
+								x="time"
+								axis="x"
+								series={[
+									{
+										key: 'used',
+										label: memoryChartConfig.used.label,
+										color: memoryChartConfig.used.color
+									},
+									{
+										key: 'process',
+										label: memoryChartConfig.process.label,
+										color: memoryChartConfig.process.color
+									}
+								]}
+								props={{
+									area: {
+										curve: curveNatural,
+										line: { class: 'stroke-1' },
+										'fill-opacity': 0.4
+									},
+									xAxis: {
+										format: () => ''
+									}
+								}}
+							>
+								{#snippet tooltip()}
+									<MyChartTooltip
+										labelFormatter={(v: Date) => {
+											return new Intl.DateTimeFormat('en-US', {
+												hour: '2-digit',
+												minute: '2-digit',
+												second: '2-digit',
+												hour12: true
+											}).format(v);
+										}}
+										valueFormatter={(v: number) => formatBytes(v)}
+										indicator="line"
+									/>
+								{/snippet}
+							</AreaChart>
+						</Chart.Container>
+					{:else}
+						<div class="text-muted-foreground flex h-[200px] items-center justify-center text-sm">
+							等待数据...
+						</div>
+					{/if}
+				</CardContent>
+			</Card>
+
+			<!-- CPU 使用情况 -->
+			<Card>
+				<CardHeader class="flex flex-row items-center justify-between space-y-0 pb-2">
+					<CardTitle class="text-sm font-medium">CPU 使用情况</CardTitle>
+					<CpuIcon class="text-muted-foreground h-4 w-4" />
+				</CardHeader>
+				<CardContent>
+					{#if sysInfo}
+						<div class="mb-4 space-y-2">
+							<div class="flex items-center justify-between text-sm">
+								<span>当前 CPU 使用率</span>
+								<span class="font-medium">{formatCpu(sysInfo.used_cpu)}</span>
+							</div>
+						</div>
+					{/if}
+					{#if cpuHistory.length > 0}
+						<Chart.Container config={cpuChartConfig} class="h-[150px] w-full">
+							<AreaChart
+								data={cpuHistory}
+								x="time"
+								axis="x"
+								series={[
+									{
+										key: 'used',
+										label: cpuChartConfig.used.label,
+										color: cpuChartConfig.used.color
+									},
+									{
+										key: 'process',
+										label: cpuChartConfig.process.label,
+										color: cpuChartConfig.process.color
+									}
+								]}
+								props={{
+									area: {
+										curve: curveNatural,
+										line: { class: 'stroke-1' },
+										'fill-opacity': 0.4
+									},
+									xAxis: {
+										format: () => ''
+									}
+								}}
+							>
+								{#snippet tooltip()}
+									<MyChartTooltip
+										labelFormatter={(v: Date) => {
+											return new Intl.DateTimeFormat('en-US', {
+												hour: '2-digit',
+												minute: '2-digit',
+												second: '2-digit',
+												hour12: true
+											}).format(v);
+										}}
+										valueFormatter={(v: number) => formatCpu(v)}
+										indicator="line"
+									/>
+								{/snippet}
+							</AreaChart>
+						</Chart.Container>
+					{:else}
+						<div class="text-muted-foreground flex h-[150px] items-center justify-center text-sm">
+							等待数据...
+						</div>
+					{/if}
+				</CardContent>
+			</Card>
+		</div>
+	{/if}
 </div>
-
-<!-- 统计信息 -->
-{#if videosData}
-	<div class="mb-6 flex items-center justify-between">
-		<div class="flex items-center gap-4">
-			<div class="text-muted-foreground text-sm">
-				共 {videosData.total_count} 个视频
-			</div>
-			<div class="text-muted-foreground text-sm">
-				共 {totalPages} 页
-			</div>
-		</div>
-		<div class="flex items-center gap-2">
-			<Button
-				size="sm"
-				variant="outline"
-				class="cursor-pointer text-xs"
-				onclick={() => (resetAllDialogOpen = true)}
-				disabled={resettingAll || loading}
-			>
-				<RotateCcwIcon class="mr-1.5 h-3 w-3 {resettingAll ? 'animate-spin' : ''}" />
-				重置所有视频
-			</Button>
-		</div>
-	</div>
-{/if}
-
-<!-- 视频卡片网格 -->
-{#if loading}
-	<div class="flex items-center justify-center py-12">
-		<div class="text-muted-foreground">加载中...</div>
-	</div>
-{:else if videosData?.videos.length}
-	<div
-		style="display: grid; grid-template-columns: repeat(auto-fill, minmax(260px, 1fr)); gap: 16px; width: 100%; max-width: none; justify-items: start;"
-	>
-		{#each videosData.videos as video (video.id)}
-			<div style="max-width: 400px; width: 100%;">
-				<VideoCard
-					{video}
-					onReset={async () => {
-						await handleResetVideo(video.id);
-					}}
-				/>
-			</div>
-		{/each}
-	</div>
-
-	<!-- 翻页组件 -->
-	<Pagination
-		currentPage={$appStateStore.currentPage}
-		{totalPages}
-		onPageChange={handlePageChange}
-	/>
-{:else}
-	<div class="flex items-center justify-center py-12">
-		<div class="space-y-2 text-center">
-			<p class="text-muted-foreground">暂无视频数据</p>
-			<p class="text-muted-foreground text-sm">尝试搜索或检查视频来源配置</p>
-		</div>
-	</div>
-{/if}
-
-<!-- 重置所有视频确认对话框 -->
-<AlertDialog.Root bind:open={resetAllDialogOpen}>
-	<AlertDialog.Content>
-		<AlertDialog.Header>
-			<AlertDialog.Title>重置所有视频</AlertDialog.Title>
-			<AlertDialog.Description>
-				此操作将重置所有视频和分页的失败状态为未下载状态，使它们在下次下载任务中重新尝试。
-				<br />
-				<strong class="text-destructive">此操作不可撤销，确定要继续吗？</strong>
-			</AlertDialog.Description>
-		</AlertDialog.Header>
-		<AlertDialog.Footer>
-			<AlertDialog.Cancel disabled={resettingAll}>取消</AlertDialog.Cancel>
-			<AlertDialog.Action
-				onclick={handleResetAllVideos}
-				disabled={resettingAll}
-				class="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-			>
-				{#if resettingAll}
-					<RotateCcwIcon class="mr-2 h-4 w-4 animate-spin" />
-					重置中...
-				{:else}
-					确认重置
-				{/if}
-			</AlertDialog.Action>
-		</AlertDialog.Footer>
-	</AlertDialog.Content>
-</AlertDialog.Root>
